solar1/solar/solar_panel_defect_detection/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
timm>=0.6.0  # For Vision Transformer
albumentations==1.3.0
opencv-python>=4.5.0
numpy>=1.21.0
pillow>=8.0.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
tqdm>=4.62.0
requests>=2.25.0
streamlit>=1.0.0
